zjmp done

+++ virtual_machine/includes/struct.h
@@ -6,7 +6,7 @@
 /*   By: paul <paul@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/05 16:13:48 by pauljull          #+#    #+#             */
-/*   Updated: 2020/05/07 08:51:13 by paul             ###   ########.fr       */
+/*   Updated: 2020/05/14 18:22:48 by paul             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ typedef struct			 s_process
 	struct s_process	*prev;
 	size_t				no;
 	size_t				pc;
-	size_t				cycle_last_live;
+	int				cycle_last_live;
 	size_t				tab_places;
 	int					registre[16];
 	char				carry;
@@ -73,7 +73,7 @@ typedef struct			s_vm
 	t_option			opt;
 	t_process			**process_list;
 	unsigned int		param[3][2];
-	size_t				cycle;
+	int				cycle;
 	size_t				nb_process;
 	size_t				nb_max_process;
 	int					cycles_to_die;
